pyproject.toml
[project]
name = "AudioBookRequest"
version = "1.3.0"
requires-python = ">= 3.12"
readme = "README.md"
# no explicit versioning unless required. the uv.lock file is for that
dependencies = [
    "aiohttp",
    "alembic",
    "argon2-cffi",
    "argon2-cffi-bindings",
    "fastapi[standard]",
    "itsdangerous",
    "jinja2",
    "jinja2-fragments",
    "pydantic",
    "pydantic-settings",
    "pydantic-core",
    "pyjwt",
    "pytailwindcss",
    "sqlmodel",
    "terminaltables",
    "torrent-parser",
    "typer",
    "urllib3",
    "rapidfuzz",
    "python-multipart",
    "requests",
    "structlog",
]

# setuptools by default expects a "src" folder structure
[tool.setuptools.packages.find]
where = ["."]

[tool.uv]
package = false

[tool.pyright]
include = ["**/*.py"]
exclude = ["**/__pycache__", "**/.venv", "**/.direnv"]
ignore = []

typeCheckingMode = "strict"
reportUnknownParameterType = true
reportMissingParameterType = true

[tool.djlint]
profile = "jinja"
ignore = "H021"

[dependency-groups]
dev = ["djlint", "pyright", "ruff"]
